requirements.txt

Django==3.1.4

djangorestframework==3.12.2
Pillow==8.0.1
ipdb==0.13.4
python-resize-image==1.1.19
psycopg2==2.8.6
django-model-utils==4.1.1

gunicorn==20.0.4
whitenoise==5.2.0

boto3==1.16.47
django-storages==1.11.1

pyyaml
ua-parser
user-agents
django-user-agents
python-gettext
jedi==0.17.2
django-cors-headers